improve miyka-recurse.sh [2]

+++ scripts/miyka-recurse.sh
@@ -3,8 +3,11 @@
 SOURCE="$1"
 TARGET="$2"
 
-if ! test "$(readlink -f "$SOURCE")" = "$(readlink -f "$TARGET")"
+if test -L "$TARGET"
 then
-	rm -rf "$SOURCE"
-	exit 1
+	if ! test "$(readlink -f "$SOURCE")" = "$(readlink -f "$TARGET")"
+	then
+		rm -rf "$SOURCE"
+		exit 1
+	fi
 fi
